hfp: add hfp_h2_sync implementation
